Create HashMap_based.c
A Hash Map Based Implementation of the Spell Checker and Recommender feature. Hash Map is great for easy lookup. Further used the Levenshtein Distance for recommendations.

Lookup avg case O(1)
Recommender  O(N.L^2 + NlogN)
where N are the number of words in dictionary and L are the number of alphabets in the words. Recommender has large time complexity for larger dictionaries